refactor: 관련 파일끼리 가까이 위치하도록 변경
